Changed Haskell program functionality
Now, Haskell will take the same arguments, but it returns something much simpler than before. It returns an entire list of new points. No point keys that will be removed. Just a list of points and the next key that will be used. In Java I just need to modify the code to redefine my Points() variable every frame so that it can read in the new points.