Added dampening factor. I think I fixed the collision issue; it was a visual thing
The dampening coefficient will soften the blow if a planet gets too close to another planet and  basically explodes when the force goes to infinity. This is done in the Haskell file. I changed line 170 of the Java program to allow for the circles to be displayed in their true positions.